+++ controllers/exclusionsController.js
@@ -2,8 +2,12 @@ import Exclusion from '../models/exclusionsModel.js'
 import Meal from '../models/mealsModel.js'
 import Diet from '../models/dietsModel.js'
 import {ApiError} from "../utils/errors.js";
+import {checkPermissions} from "../utils/auth.js";
 
 export const getExclusions = async (req, res, next) => {
+    if(!checkPermissions(req.userInfo, process.env.ACCESS_USER)){
+        return next(ApiError("You're not authorized to perform this action!", 401))
+    }
     try {
         const exclusions = await Exclusion.find({});
         res.status(200);
@@ -14,6 +18,9 @@ export const getExclusions = async (req, res, next) => {
     }
 }
 export const addExclusion = async (req, res, next) => {
+    if(!checkPermissions(req.userInfo, process.env.ACCESS_DIETETICIAN)){
+        return next(ApiError("You're not authorized to perform this action!", 401))
+    }
     const exclusionName = req.body.name;
     try {
         const existingExclusion = await Exclusion.findOne({name: exclusionName})
@@ -35,6 +42,9 @@ export const addExclusion = async (req, res, next) => {
     }
 }
 export const updateExclusion = async (req, res, next) => {
+    if(!checkPermissions(req.userInfo, process.env.ACCESS_DIETETICIAN)){
+        return next(ApiError("You're not authorized to perform this action!", 401))
+    }
     const exclusionName = req.body.name;
     const id = req.params.id
     try {
@@ -50,6 +60,9 @@ export const updateExclusion = async (req, res, next) => {
     }
 }
 export const deleteExclusion = async (req, res, next) => {
+    if(!checkPermissions(req.userInfo, process.env.ACCESS_DIETETICIAN)){
+        return next(ApiError("You're not authorized to perform this action!", 401))
+    }
     const id = req.params.id
     try {
         const deletedExclusion = await Exclusion.findByIdAndDelete(id);
